+++ packages/entity-graph-sdl/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @prometheus-ags/entity-graph-sdl
 
+## 3.0.5
+
 ## 3.0.4
 
 ## 3.0.3