+++ packages/entity-graph-sync/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prometheus-ags/entity-graph-sync
 
+## 3.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @prometheus-ags/entity-graph-core@3.0.5
+
 ## 3.0.4
 
 ### Patch Changes